Add ColumnReference type with new API (#3)
Previously cell references were stringly-typed allowing a user to create a cell reference with an arbitrary string. This is fixed with an introduction of new types: `CellReference` and `ColumnReference`. `ColumnReference` is a simple wrapper around `String` that guards against empty string and characters that can't be used as column references. Rows don't need a separate type for references as they map neatly to `UInt`, although API parts that used row references have changed to use `UInt` instead of `Int`. This is a breaking change that requires a major version bump.