Add SharedStrings model, parse sharedStrings.xml (#8)
As reported in #21, there's no obvious way to get values for some values. That's caused by the fact that some of the values are stored in a separate file, which is usually located at the path `sharedStrings.xml`. This can be parsed to a corresponding model type and then processed by users of `CoreXLSX` to get the string values.

* Add SharedStrings model, general cleanup
* Add link to Open XML SDK docs for SharedStrings
* Fix rebase conflicts
* Add public `parseSharedStrings` API with tests